Fix to use without any type
