+++ src/core/createContext.tsx
@@ -1,23 +1,13 @@
 import React, { createContext, createElement } from 'react';
-import { Store } from './store';
 import { Subscription } from './subscription';
 import { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect';
 
-export type ContextProvider<T extends Record<string, unknown>> = {
-  children: React.ReactNode;
-  store?: Store<T>;
-};
-
-export type ContextSource = {
-  [displayName: string]: any;
-};
-
-const dispatchEventLister = <T extends ContextSource>(
-  provider: React.Context<any>['Provider'],
-  subscription: Subscription<T>
+const dispatchEventLister = <State extends any>(
+  provider: React.Context<State>['Provider'],
+  subscription: Subscription<State>
 ) => {
   const dispatcher = React.memo(
-    ({ value, children }: React.ProviderProps<any>) => {
+    ({ value, children }: React.ProviderProps<State>) => {
       useIsomorphicLayoutEffect(() => {
         subscription.forEach((listener) => {
           listener(value);
@@ -28,19 +18,22 @@ const dispatchEventLister = <T extends ContextSource>(
     }
   );
 
+  dispatcher.displayName = 'UseGlobalContextProvider';
+
   return dispatcher;
 };
 
-export const createBaseContext = <T extends ContextSource>() => {
-  const stateContext = createContext(null as any, () => 0);
-  const dispatchContext = createContext(null as any);
+export const createBaseContext = <State, Dispatch>() => {
+  const stateContext = createContext<State>(null as any, () => 0);
+  const dispatchContext = createContext<Dispatch>(null as any);
   const subscription = new Set<() => void>();
 
-  stateContext.Provider = dispatchEventLister<T>(
+  stateContext.Provider = dispatchEventLister<State>(
     stateContext.Provider,
     subscription
   );
   stateContext.displayName = 'UseGlobalStateContext';
+  dispatchContext.displayName = 'UseGlobalDispatchContext';
 
   return { stateContext, dispatchContext, subscription };
 };
